requirements.txt
streamlit==1.29.0
snowflake-connector-python==3.5.0
pandas==2.1.4
urllib3<2.0.0